# Optimise `set.Union`