## Summary

<!-- Describe your changes in detail here, if it closes an open issue, include "Closes #<issue>" -->
Did some profiling and benchmarking and it turns out that allocation is the biggest time cost in `set.Union`, so after some experimentation and measurement it turns out it's actually faster to iterate through the sets *twice*; once to calculate the required capacity for the union, and the second time to fill it.

Doing so we reduce both the time taken and the memory allocations significantly as the set doesn't have to constantly grow:

```plaintext
goos: darwin
goarch: amd64
pkg: github.com/FollowTheProcess/collections/set
cpu: Intel(R) Core(TM) i5-8257U CPU @ 1.40GHz
        │  before.txt  │              after.txt              │
        │    sec/op    │   sec/op     vs base                │
Union-8   117.30µ ± 1%   78.71µ ± 1%  -32.89% (p=0.000 n=30)

        │  before.txt  │              after.txt              │
        │     B/op     │     B/op      vs base               │
Union-8   48.92Ki ± 0%   48.07Ki ± 0%  -1.74% (p=0.000 n=30)

        │ before.txt  │             after.txt              │
        │  allocs/op  │ allocs/op   vs base                │
Union-8   88.000 ± 0%   4.000 ± 0%  -95.45% (p=0.000 n=30)
```